pass interfaces to card component
