+++ components/Card.vue
@@ -1,5 +1,5 @@
 <template>
-  <svg class="NuxtLogo" width="245" height="180" viewBox="0 0 452 342" xmlns="http://www.w3.org/2000/svg">
+  <svg width="245" height="180" viewBox="0 0 452 342" xmlns="http://www.w3.org/2000/svg">
     <g fill="none" fill-rule="evenodd">
       <path
         d="M139 330l-1-2c-2-4-2-8-1-13H29L189 31l67 121 22-16-67-121c-1-2-9-14-22-14-6 0-15 2-22 15L5 303c-1 3-8 16-2 27 4 6 10 12 24 12h136c-14 0-21-6-24-12z"
@@ -17,18 +17,38 @@
     </g>
   </svg>
 </template>
-<style>
-.NuxtLogo {
-  animation: 1s appear;
-  margin: auto;
-}
 
-@keyframes appear {
-  0% {
-    opacity: 0;
+<script lang="ts">
+import { Component, Vue } from 'nuxt-property-decorator'
+
+export interface Card {
+    id: String
+    company_id: String
+    name: String
+    thumb_url: String | null
+    priority: Number,
+    due_date: DueDate | null
+    status: String
+    enrollment?: Enrollment
   }
-  100% {
-    opacity: 1;
+
+export interface DueDate {
+    available_at: Number
   }
+
+export interface Enrollment {
+    id: String
+    percentage: Number
+  }
+
+export enum Status {
+    IN_PROGRESS = 'IN_PROGRESS',
+    NOT_STARTED = 'NOT_STARTED',
+    COMPLETED = 'COMPLETED',
+    BLOCKED = 'BLOCKED'
+  }
+
+  @Component({ name: 'VCard' })
+export default class VCard extends Vue {
 }
-</style>
+</script>
